gps: Move user details view in gps module
It's a GPS beneficiary view with the follow up group members, and now
the gps related actions.
No reasons to keep it in a different module
