+++ config/urls.py
@@ -84,7 +84,6 @@ urlpatterns = [
     path("logout/", include("itou.www.logout.urls")),
     path("signup/", include("itou.www.signup.urls")),
     path("stats/", include("itou.www.stats.urls")),
-    path("users/", include("itou.www.users_views.urls")),
     path("announcements/", include("itou.www.announcements.urls")),
     path("versions/", include("itou.www.releases.urls")),
     # Enable Mailjet status tracking
